+++ .github/workflows/preconditions.yml
@@ -36,9 +36,12 @@ jobs:
           cache: 'maven'
       - name: Setup Go
         uses: actions/setup-go@44e221478fc6847752e5c574fc7a7b3247b00fbf
+        with:
+          go-version: '1.17'
       - name: Get Changed Files
+        id: changed-files
         uses: tj-actions/changed-files@61ee456a9d0f512e7ecfdf28863634c97dae2d16
         with:
           separator: ','
       - name: Run Spotless
-        run: go run ./cicd/cmd/run-spotless.go --changed-files="${{ steps.changed-files.outputs.all_changed_and_modified_files }}"
+        run: go run ./cicd/cmd/run-spotless.go --changed-files="${{ steps.changed-files.outputs.all_changed_files }}"
