+++ cicd/go.mod
@@ -1,0 +1,3 @@
+module github.com/GoogleCloudPlatform/DataflowTemplates/cicd
+
+go 1.17
